requirements.txt
# Core Dependencies
opencv-python>=4.7.0
ultralytics>=8.0.0
psycopg2-binary>=2.9.5
PyQt5>=5.15.9
pandas>=1.5.3
matplotlib>=3.7.1
win10toast>=0.9
python-dotenv>=1.0.0
playsound>=1.2.2
deepface>=0.0.79
tensorflow>=2.13.0
scikit-learn>=1.3.0

# Development
pytest>=7.2.0
black>=22.12.0
flake8>=6.0.0
isort>=5.12.0